+++ lib/rage/code_loader.rb
@@ -37,6 +37,7 @@ class Rage::CodeLoader
     load("#{Rage.root}/config/routes.rb")
 
     reload_components
+    Rage.config.run_hooks_for(:after_reload)
   end
 
   # in Rails mode - reset the routes; everything else will be done by Rails
@@ -47,6 +48,7 @@ class Rage::CodeLoader
     Rage.__router.reset_routes
 
     reload_components
+    Rage.config.run_hooks_for(:after_reload)
   end
 
   def reloading?
@@ -54,7 +56,9 @@ class Rage::CodeLoader
   end
 
   def check_updated!
-    current_watched = @autoload_path.glob("**/*.rb") + Rage.root.glob("config/routes.rb") + Rage.root.glob("config/openapi_components.*")
+    current_watched = @autoload_path.glob("**/*.{rb,erb}") + Rage.root.glob("config/routes.rb") + Rage.root.glob("config/openapi_components.*")
+    current_watched += Rage.config.code_loader.reload_paths.flat_map { |pattern| Rage.root.glob(pattern) }
+
     current_update_at = current_watched.max_by { |path| path.exist? ? path.mtime.to_f : 0 }&.mtime.to_f
     return false if !@last_watched && !@last_update_at