+++ lib/rage/configuration.rb
@@ -141,6 +141,15 @@ class Rage::Configuration
     push_hook(block, :after_initialize)
   end
 
+  # Schedule a block of code to run after Rage has reloaded the application code in development. Use this to reset state or re-initialize dependencies that cache application-level constants.
+  # @example
+  #   Rage.config.after_reload do
+  #     MyCache.clear
+  #   end
+  def after_reload(&block)
+    push_hook(block, :after_reload)
+  end
+
   # Register a custom renderer that generates overloads `render` on all controllers.
   # The block receives the object passed to `render` together with any additional keyword arguments.
   # The code inside the block is executed in the context of the controller instance, so you can access all usual controller methods in it.
@@ -302,6 +311,14 @@ class Rage::Configuration
   # @return [Rage::Configuration::Router]
   def router
     @router ||= Router.new
+  end
+  # @!endgroup
+
+  # @!group Code Loader Configuration
+  # Allows configuring code loader settings.
+  # @return [Rage::Configuration::CodeLoader]
+  def code_loader
+    @code_loader ||= CodeLoader.new
   end
   # @!endgroup
 
@@ -1145,6 +1162,29 @@ class Rage::Configuration
     #       config.router.form_actions = true
     #     end
     attr_accessor :form_actions
+  end
+
+  class CodeLoader
+    # @private
+    def initialize
+      @reload_paths = []
+    end
+
+    # @private
+    attr_reader :reload_paths
+
+    # Specify additional paths to watch for changes in development.
+    #
+    # @param paths [Array<Pathname, String>] glob patterns or directory paths to watch
+    # @example Watch HAML templates
+    #   Rage.configure do
+    #     config.code_loader.reload_paths = ["app/views/**/*.haml"]
+    #   end
+    def reload_paths=(paths)
+      @reload_paths = Array(paths).map do |path|
+        path.is_a?(Pathname) ? path : Pathname.new(path)
+      end
+    end
   end
 
   class BlockingOperationPool